StudyFlow/requirements.txt
pyautogui
pytesseract
Pillow
openai
anthropic
cohere
opencv-python
numpy
python-dotenv
tk
Flask
gunicorn
